requirements.txt
audioop-lts==0.2.2
beautifulsoup4==4.14.2
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.1.8
gTTS==2.5.4
idna==3.11
PyAudio==0.2.14
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.5
six==1.17.0
soupsieve==2.8
SpeechRecognition==3.14.3
standard-aifc==3.13.0
standard-chunk==3.13.0
typing_extensions==4.15.0
urllib3==2.5.0
wikipedia==1.4.0
